requirements.txt
Django==2.2.5
python-telegram-bot==12.1.1
django-environ
psycopg2-binary
django-extensions